+++ CHANGELOG.md
@@ -10,7 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- New module: `I18n` (*Internationalization*), for translating string
+  to the user's language. Configuration files for 15 languages.
 - `Timeout.add_ignore`
+- `Widget.remove_connection`, `Widget.remove_trigger`
 
 ## [20250219](https://github.com/sanette/bogue/compare/20241212...20250219)